feat(clients): remove legacy broad-client bridge helpers (task 2)
Delete ConvertProviderData, MaybeNewAPIClientFromFrameworkResource,
MaybeNewKibanaAPIClientFromFrameworkResource, NewAPIClientFromSDKResource,
NewKibanaAPIClientFromSDKResource, and extractDefaultClientFromMeta from
api_client.go — all production callers were already migrated in Task 1.

Remove GetDefaultClient() from ProviderClientFactory; it was only needed
by the now-deleted bridge helpers.

Move shared test helpers (newTestAPIClient, newMockKibanaServer,
kibanaConnectionAttrTypes) into test_helpers_test.go so they remain
available to provider_client_factory_test.go and
elasticsearch_scoped_client_test.go. Delete api_client_kibana_test.go
whose tests exclusively exercised the removed functions.
