+++ internal/clients/api_client.go
@@ -36,7 +36,6 @@ import (
 	"github.com/elastic/terraform-provider-elasticstack/internal/models"
 	"github.com/hashicorp/go-version"
 	fwdiags "github.com/hashicorp/terraform-plugin-framework/diag"
-	"github.com/hashicorp/terraform-plugin-framework/types"
 	"github.com/hashicorp/terraform-plugin-log/tflog"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/logging"
@@ -158,221 +157,6 @@ func NewAPIClientFromFramework(ctx context.Context, cfg config.ProviderConfigura
 	}
 
 	return client, nil
-}
-
-func ConvertProviderData(providerData any) (*APIClient, fwdiags.Diagnostics) {
-	var diags fwdiags.Diagnostics
-
-	if providerData == nil {
-		return nil, diags
-	}
-
-	// Support the new factory injection path: extract the default client.
-	if factory, ok := providerData.(*ProviderClientFactory); ok {
-		if factory == nil {
-			diags.AddError(
-				"Unconfigured Client Factory",
-				"Expected configured client factory. Please report this issue to the provider developers.",
-			)
-			return nil, diags
-		}
-		client := factory.GetDefaultClient()
-		if client == nil {
-			diags.AddError(
-				"Unconfigured Client",
-				"Expected configured client. Please report this issue to the provider developers.",
-			)
-			return nil, diags
-		}
-		return client, diags
-	}
-
-	client, ok := providerData.(*APIClient)
-	if !ok {
-		diags.AddError(
-			"Unexpected Provider Data",
-			fmt.Sprintf("Expected *ProviderClientFactory or *APIClient, got: %T. Please report this issue to the provider developers.", providerData),
-		)
-
-		return nil, diags
-	}
-	if client == nil {
-		diags.AddError(
-			"Unconfigured Client",
-			"Expected configured client. Please report this issue to the provider developers.",
-		)
-	}
-	return client, diags
-}
-
-func MaybeNewAPIClientFromFrameworkResource(ctx context.Context, esConnList types.List, defaultClient *APIClient) (*APIClient, fwdiags.Diagnostics) {
-	var esConns []config.ElasticsearchConnection
-	if diags := esConnList.ElementsAs(ctx, &esConns, true); diags.HasError() {
-		return nil, diags
-	}
-
-	if len(esConns) == 0 {
-		return defaultClient, nil
-	}
-
-	cfg, diags := config.NewFromFramework(ctx, config.ProviderConfiguration{Elasticsearch: esConns}, defaultClient.version)
-	if diags.HasError() {
-		return nil, diags
-	}
-
-	esClient, err := buildEsClient(cfg)
-	if err != nil {
-		return nil, fwdiags.Diagnostics{fwdiags.NewErrorDiagnostic(err.Error(), err.Error())}
-	}
-
-	return &APIClient{
-		elasticsearch: esClient,
-		kibana:        defaultClient.kibana,
-		fleet:         defaultClient.fleet,
-		version:       defaultClient.version,
-	}, diags
-}
-
-// MaybeNewKibanaAPIClientFromFrameworkResource resolves a scoped *APIClient
-// from an entity-local kibana_connection block. If the block is absent the
-// default client is returned unchanged. If the block is configured a new
-// *APIClient is returned with the Kibana legacy, Kibana OpenAPI, SLO, and
-// Fleet clients rebuilt from the scoped connection. The resulting client does
-// not carry a provider-level Elasticsearch client so that version and identity
-// checks resolve against the scoped Kibana connection.
-func MaybeNewKibanaAPIClientFromFrameworkResource(ctx context.Context, kibConnList types.List, defaultClient *APIClient) (*APIClient, fwdiags.Diagnostics) {
-	var kibConns []config.KibanaConnection
-	if diags := kibConnList.ElementsAs(ctx, &kibConns, true); diags.HasError() {
-		return nil, diags
-	}
-
-	if len(kibConns) == 0 {
-		return defaultClient, nil
-	}
-
-	cfg, diags := config.NewFromFrameworkKibanaResource(ctx, kibConns, defaultClient.version)
-	if diags.HasError() {
-		return nil, diags
-	}
-
-	kibanaClient, err := buildKibanaClient(*cfg)
-	if err != nil {
-		return nil, fwdiags.Diagnostics{fwdiags.NewErrorDiagnostic("Failed to build Kibana client", err.Error())}
-	}
-
-	kibanaOapiClient, err := buildKibanaOapiClient(*cfg)
-	if err != nil {
-		return nil, fwdiags.Diagnostics{fwdiags.NewErrorDiagnostic("Failed to build Kibana OpenAPI client", err.Error())}
-	}
-
-	fleetClient, err := buildFleetClient(*cfg)
-	if err != nil {
-		return nil, fwdiags.Diagnostics{fwdiags.NewErrorDiagnostic("Failed to build Fleet client", err.Error())}
-	}
-
-	var sloAPI slo.SloAPI
-	if kibanaClient != nil {
-		kibanaHTTPClient := kibanaClient.Client.GetClient()
-		sloAPI = buildSloClient(*cfg, kibanaHTTPClient).SloAPI
-	}
-
-	return &APIClient{
-		kibana:       kibanaClient,
-		kibanaOapi:   kibanaOapiClient,
-		slo:          sloAPI,
-		kibanaConfig: *cfg.Kibana,
-		fleet:        fleetClient,
-		version:      defaultClient.version,
-	}, nil
-}
-
-// NewKibanaAPIClientFromSDKResource resolves a scoped *APIClient from an
-// entity-local kibana_connection block in a Plugin SDK resource. If the block
-// is absent the provider-level default client is returned unchanged. If the
-// block is configured a new *APIClient is returned with all Kibana-derived
-// clients rebuilt from the scoped connection.
-func NewKibanaAPIClientFromSDKResource(d *schema.ResourceData, meta any) (*APIClient, diag.Diagnostics) {
-	defaultClient := extractDefaultClientFromMeta(meta)
-	version := defaultClient.version
-
-	resourceConfig, diags := config.NewFromSDKKibanaResource(d, version)
-	if diags.HasError() {
-		return nil, diags
-	}
-
-	if resourceConfig == nil {
-		return defaultClient, nil
-	}
-
-	kibanaClient, err := buildKibanaClient(*resourceConfig)
-	if err != nil {
-		return nil, diag.FromErr(err)
-	}
-
-	kibanaOapiClient, err := buildKibanaOapiClient(*resourceConfig)
-	if err != nil {
-		return nil, diag.FromErr(err)
-	}
-
-	fleetClient, err := buildFleetClient(*resourceConfig)
-	if err != nil {
-		return nil, diag.FromErr(err)
-	}
-
-	var sloAPI slo.SloAPI
-	if kibanaClient != nil {
-		kibanaHTTPClient := kibanaClient.Client.GetClient()
-		sloAPI = buildSloClient(*resourceConfig, kibanaHTTPClient).SloAPI
-	}
-
-	return &APIClient{
-		kibana:       kibanaClient,
-		kibanaOapi:   kibanaOapiClient,
-		slo:          sloAPI,
-		kibanaConfig: *resourceConfig.Kibana,
-		fleet:        fleetClient,
-		version:      version,
-	}, nil
-}
-
-func NewAPIClientFromSDKResource(d *schema.ResourceData, meta any) (*APIClient, diag.Diagnostics) {
-	defaultClient := extractDefaultClientFromMeta(meta)
-	version := defaultClient.version
-	resourceConfig, diags := config.NewFromSDKResource(d, version)
-	if diags.HasError() {
-		return nil, diags
-	}
-
-	if resourceConfig == nil {
-		return defaultClient, nil
-	}
-
-	esClient, err := buildEsClient(*resourceConfig)
-	if err != nil {
-		return nil, diag.FromErr(err)
-	}
-
-	return &APIClient{
-		elasticsearch:            esClient,
-		elasticsearchClusterInfo: defaultClient.elasticsearchClusterInfo,
-		kibana:                   defaultClient.kibana,
-		fleet:                    defaultClient.fleet,
-		version:                  version,
-	}, diags
-}
-
-// extractDefaultClientFromMeta extracts the provider-level default *APIClient
-// from the SDK meta value. It handles both the legacy *APIClient case (used in
-// tests that bypass the factory) and the current *ProviderClientFactory case.
-func extractDefaultClientFromMeta(meta any) *APIClient {
-	if factory, ok := meta.(*ProviderClientFactory); ok {
-		return factory.GetDefaultClient()
-	}
-	// Legacy path: meta is a bare *APIClient (e.g. in unit tests).
-	if client, ok := meta.(*APIClient); ok {
-		return client
-	}
-	panic(fmt.Sprintf("extractDefaultClientFromMeta: unsupported meta type %T", meta))
 }
 
 func (a *APIClient) GetESClient() (*elasticsearch.Client, error) {
